[WIP] Fix test
