+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - [REFACTOR] Fix gradle warnings (Deprecated API, PredefinedEnumMap)
 - [REFACTOR] Replace some clickable to card function + reuse orange app bar
 - [REFACTOR] Google Compose Navigation (archive, firstpair, widgets, updater, import)
+- [REFACTOR] Migrate to new device info api
 
 # 1.4.1
