+++ components/bridge/rpcinfo/api/build.gradle.kts
@@ -4,6 +4,8 @@ plugins {
 
 dependencies {
     implementation(projects.components.core.data)
+    implementation(projects.components.core.ktx)
+
 
     implementation(projects.components.bridge.api)
     implementation(projects.components.bridge.service.api)
