+++ components/bridge/rpcinfo/api/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="info_device_info_flash_not_found">Not Found</string>
+</resources>
