+++ components/bridge/rpcinfo/impl/build.gradle.kts
@@ -26,4 +26,14 @@ dependencies {
     implementation(libs.dagger)
     kapt(libs.dagger.kapt)
 
+
+    // Testing
+    testImplementation(projects.components.core.test)
+    testImplementation(libs.junit)
+    testImplementation(libs.mockk)
+    testImplementation(libs.mockito.kotlin)
+    testImplementation(libs.ktx.testing)
+    testImplementation(libs.roboelectric)
+    testImplementation(libs.lifecycle.test)
+    testImplementation(libs.kotlin.coroutines.test)
 }
