+++ components/bridge/rpcinfo/impl/src/test/java/com/flipperdevices/bridge/rpcinfo/impl/mapper/DeviceInfoHelperTest.kt
@@ -1,13 +1,11 @@
-package com.flipperdevices.bridge.impl.manager.service
+package com.flipperdevices.bridge.rpcinfo.impl.mapper
 
-import com.flipperdevices.bridge.api.model.FirmwareInfo
-import com.flipperdevices.bridge.api.model.FlipperDeviceInfo
-import com.flipperdevices.bridge.api.model.FlipperRpcInformation
-import com.flipperdevices.bridge.api.model.RadioStackInfo
-import com.flipperdevices.bridge.api.model.RadioStackType
+import com.flipperdevices.bridge.rpcinfo.model.FirmwareInfo
+import com.flipperdevices.bridge.rpcinfo.model.FlipperDeviceInfo
+import com.flipperdevices.bridge.rpcinfo.model.FlipperRpcInformation
+import com.flipperdevices.bridge.rpcinfo.model.RadioStackInfo
+import com.flipperdevices.bridge.rpcinfo.model.RadioStackType
 import com.flipperdevices.core.data.SemVer
-import com.flipperdevices.info.impl.viewmodel.helper.DeviceInfoHelper
-import com.flipperdevices.info.impl.viewmodel.helper.InternalFlipperRpcInformationRaw
 import kotlinx.collections.immutable.persistentMapOf
 import org.junit.Assert
 import org.junit.Test
@@ -21,12 +19,11 @@ class DeviceInfoHelperTest(
 ) {
     @Test
     fun `Parse fields`() {
-        val deviceInfo =
-            com.flipperdevices.info.impl.viewmodel.helper.DeviceInfoHelper.mapRawRpcInformation(
-                com.flipperdevices.info.impl.viewmodel.helper.InternalFlipperRpcInformationRaw(
-                    otherFields = fields
-                )
+        val deviceInfo = DeprecatedFlipperRpcInfoMapper().map(
+            InternalFlipperRpcInformationRaw(
+                otherFields = fields
             )
+        )
         Assert.assertEquals(infoExpected.flipperDeviceInfo, deviceInfo.flipperDeviceInfo)
         Assert.assertEquals(infoExpected.firmware, deviceInfo.firmware)
         Assert.assertEquals(infoExpected.radioStack, deviceInfo.radioStack)
