+++ components/info/impl/src/main/res/values/strings.xml
@@ -36,7 +36,6 @@
     <string name="info_device_info_build_date">Build Date</string>
     <string name="info_device_info_int_flash">Int. Flash (Used/Total)</string>
     <string name="info_device_info_ext_flash">SD Card (Used/Total)</string>
-    <string name="info_device_info_flash_not_found">Not Found</string>
     <string name="info_device_info_more_information">Full Info</string>
 
     <string name="full_info_flipper_device">Flipper Device</string>
